backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.7.1
numpy>=1.24.3
torch>=2.1.0
transformers>=4.34.0
Pillow>=10.0.1
opencv-python>=4.8.1.78
librosa>=0.10.1
requests>=2.31.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
sqlalchemy>=2.0.22
alembic>=1.12.1
pytest>=7.4.3
httpx>=0.25.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.1 
pydantic-settings>=2.0.0
# For fine-tuning
datasets>=2.16.0
torchvision>=0.15.0